pruner: move accessors_test.go to pruner_test package
