+++ pruner/accessors_test.go
@@ -1,150 +1,19 @@
-package pruner
+package pruner_test
 
 import (
 	"context"
 	"testing"
 
 	"github.com/NethermindEth/juno/core"
-	"github.com/NethermindEth/juno/core/felt"
 	"github.com/NethermindEth/juno/db"
 	_ "github.com/NethermindEth/juno/encoder/registry"
+	"github.com/NethermindEth/juno/pruner"
 	"github.com/NethermindEth/juno/pruner/testutils"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
-func TestOldestRetainedBlock(t *testing.T) {
-	t.Run("empty database returns ErrKeyNotFound", func(t *testing.T) {
-		database := testutils.NewPebbleTestDB(t)
-
-		_, err := OldestRetainedBlock(database)
-		assert.ErrorIs(t, err, db.ErrKeyNotFound)
-	})
-
-	t.Run("returns lowest block number with commitments", func(t *testing.T) {
-		database := testutils.NewPebbleTestDB(t)
-
-		for i := uint64(5); i <= 7; i++ {
-			testutils.StoreBlock(t, database, i)
-		}
-
-		num, err := OldestRetainedBlock(database)
-		require.NoError(t, err)
-		assert.Equal(t, uint64(5), num)
-	})
-}
-
-func TestDeleteTransactionHashReverseLookups(t *testing.T) {
-	database := testutils.NewPebbleTestDB(t)
-
-	blocks := make([]*testutils.StoredBlock, 3)
-	for i := range uint64(3) {
-		blocks[i] = testutils.StoreBlock(t, database, i)
-	}
-
-	testutils.WithBatch(t, database, func(batch db.Batch) error {
-		return deleteTransactionHashReverseLookups(database, batch, 1)
-	})
-
-	// Block 1's tx hash reverse lookups (bucket 9) and L1 handler msg hash
-	// (bucket 24) should be gone.
-	for _, txHash := range blocks[1].TxHashes {
-		_, err := core.TransactionBlockNumbersAndIndicesByHashBucket.Get(
-			database,
-			(*felt.TransactionHash)(txHash),
-		)
-		assert.Error(t, err)
-	}
-	for _, msgHash := range blocks[1].L1HandlerMsgHashes {
-		_, err := core.GetL1HandlerTxnHashByMsgHash(database, msgHash)
-		assert.Error(t, err)
-	}
-
-	// Header hash → number lookup is *not* this function's responsibility,
-	// so it must still be present for block 1.
-	assert.True(t, testutils.BlockHeaderHashExists(database, blocks[1].Header.Hash))
-
-	// Neighbouring blocks must be untouched.
-	for _, b := range []*testutils.StoredBlock{blocks[0], blocks[2]} {
-		for _, txHash := range b.TxHashes {
-			_, err := core.TransactionBlockNumbersAndIndicesByHashBucket.Get(
-				database,
-				(*felt.TransactionHash)(txHash),
-			)
-			assert.NoError(t, err)
-		}
-		for _, msgHash := range b.L1HandlerMsgHashes {
-			_, err := core.GetL1HandlerTxnHashByMsgHash(database, msgHash)
-			assert.NoError(t, err)
-		}
-	}
-}
-
-func TestPruneAggregatedBloomFiltersUpto(t *testing.T) {
-	n := core.NumBlocksPerFilter
-
-	writeFilters := func(t *testing.T, database db.KeyValueWriter) {
-		t.Helper()
-		for i := range uint64(3) {
-			filter := core.NewAggregatedFilter(i * n)
-			require.NoError(t, core.WriteAggregatedBloomFilter(database, &filter))
-		}
-	}
-
-	bloomFilterExists := func(database db.KeyValueReader, from, to uint64) bool {
-		_, err := core.GetAggregatedBloomFilter(database, from, to)
-		return err == nil
-	}
-
-	t.Run("deletes filters fully below the boundary", func(t *testing.T) {
-		database := testutils.NewPebbleTestDB(t)
-		writeFilters(t, database)
-
-		testutils.WithBatch(t, database, func(batch db.Batch) error {
-			return pruneAggregatedBloomFiltersUpto(batch, 2*n)
-		})
-
-		assert.False(t, bloomFilterExists(database, 0, n-1))
-		assert.False(t, bloomFilterExists(database, n, 2*n-1))
-		assert.True(t, bloomFilterExists(database, 2*n, 3*n-1))
-	})
-
-	t.Run("partial range does not delete the containing filter", func(t *testing.T) {
-		database := testutils.NewPebbleTestDB(t)
-		writeFilters(t, database)
-
-		testutils.WithBatch(t, database, func(batch db.Batch) error {
-			return pruneAggregatedBloomFiltersUpto(batch, n/2)
-		})
-
-		assert.True(t, bloomFilterExists(database, 0, n-1))
-		assert.True(t, bloomFilterExists(database, n, 2*n-1))
-		assert.True(t, bloomFilterExists(database, 2*n, 3*n-1))
-	})
-
-	t.Run("rangeEndExclusive at filter boundary keeps that filter", func(t *testing.T) {
-		// rangeEndExclusive = n means filter [0, n-1] is fully below
-		// (to = n-1 < n) and filter [n, 2n-1] starts at n, so it's kept.
-		database := testutils.NewPebbleTestDB(t)
-		writeFilters(t, database)
-
-		testutils.WithBatch(t, database, func(batch db.Batch) error {
-			return pruneAggregatedBloomFiltersUpto(batch, n)
-		})
-
-		assert.False(t, bloomFilterExists(database, 0, n-1))
-		assert.True(t, bloomFilterExists(database, n, 2*n-1))
-		assert.True(t, bloomFilterExists(database, 2*n, 3*n-1))
-	})
-
-	t.Run("no-op when range below one filter span", func(t *testing.T) {
-		database := testutils.NewPebbleTestDB(t)
-
-		testutils.WithBatch(t, database, func(batch db.Batch) error {
-			return pruneAggregatedBloomFiltersUpto(batch, 0)
-		})
-	})
-}
+const testTargetBatchByteSize = 96 * db.Megabyte
 
 // TestPruneUpto verifies the complete prune lifecycle with realistic block
 // counts: what survives, what gets deleted, and the two intentional
@@ -172,11 +41,11 @@ func TestPruneUpto(t *testing.T) {
 	for i := range totalBlocks {
 		blocks[i] = testutils.StoreBlock(t, database, i)
 	}
-	pruned, oldestKept, err := PruneUpto(
+	pruned, oldestKept, err := pruner.PruneUpto(
 		t.Context(),
 		database,
 		endExclusive,
-		defaultTargetBatchByteSize,
+		testTargetBatchByteSize,
 	)
 	require.NoError(t, err)
 	assert.Equal(t, endExclusive, pruned)
@@ -188,7 +57,7 @@ func TestPruneUpto(t *testing.T) {
 func TestPruneUpto_NoOp(t *testing.T) {
 	t.Run("empty database returns 0", func(t *testing.T) {
 		database := testutils.NewPebbleTestDB(t)
-		pruned, oldestKept, err := PruneUpto(t.Context(), database, 20, defaultTargetBatchByteSize)
+		pruned, oldestKept, err := pruner.PruneUpto(t.Context(), database, 20, testTargetBatchByteSize)
 		require.NoError(t, err)
 		assert.Zero(t, pruned)
 		assert.Zero(t, oldestKept)
@@ -210,11 +79,11 @@ func TestPruneUpto_NoOp(t *testing.T) {
 			for i := tc.chainStart; i < 30; i++ {
 				blocks[i] = testutils.StoreBlock(t, database, i)
 			}
-			pruned, oldestKept, err := PruneUpto(
+			pruned, oldestKept, err := pruner.PruneUpto(
 				t.Context(),
 				database,
 				tc.endExclusive,
-				defaultTargetBatchByteSize,
+				testTargetBatchByteSize,
 			)
 			require.NoError(t, err)
 			assert.Zero(t, pruned)
@@ -263,14 +132,14 @@ func TestPruneUpto_ResumeAfterMidwayCancel(t *testing.T) {
 	ctx, cancel := context.WithCancel(t.Context())
 	wrapped := &cancelAfterGets{KeyValueStore: database, cancel: cancel, after: cancelAfter}
 
-	pruned1, oldestKept1, err := PruneUpto(ctx, wrapped, endExclusive, defaultTargetBatchByteSize)
+	pruned1, oldestKept1, err := pruner.PruneUpto(ctx, wrapped, endExclusive, testTargetBatchByteSize)
 	require.NoError(t, err)
 	require.Greater(t, pruned1, uint64(0), "first call should make progress before cancel")
 	require.Less(t, pruned1, endExclusive, "first call must not finish the full window")
 	assert.Equal(t, pruned1, oldestKept1, "oldestKept after partial prune == pruned (start was 0)")
 
-	pruned2, oldestKept2, err := PruneUpto(
-		t.Context(), database, endExclusive, defaultTargetBatchByteSize,
+	pruned2, oldestKept2, err := pruner.PruneUpto(
+		t.Context(), database, endExclusive, testTargetBatchByteSize,
 	)
 	require.NoError(t, err)
 	assert.Equal(t, endExclusive, oldestKept2)
@@ -303,14 +172,14 @@ func TestPruneUpto_RollingCarveOut(t *testing.T) {
 	}
 
 	// Call 1: prune up to 20.
-	pruned, oldestKept, err := PruneUpto(t.Context(), database, 20, defaultTargetBatchByteSize)
+	pruned, oldestKept, err := pruner.PruneUpto(t.Context(), database, 20, testTargetBatchByteSize)
 	require.NoError(t, err)
 	assert.Equal(t, uint64(20), pruned)
 	assert.Equal(t, uint64(20), oldestKept)
 	testutils.AssertPostPruneState(t, database, blocks, 20, lag)
 
 	// Call 2: prune up to 30. OldestRetainedBlock returns 20 now, so we prune 10 more.
-	pruned, oldestKept, err = PruneUpto(context.Background(), database, 30, defaultTargetBatchByteSize)
+	pruned, oldestKept, err = pruner.PruneUpto(t.Context(), database, 30, testTargetBatchByteSize)
 	require.NoError(t, err)
 	assert.Equal(t, uint64(10), pruned)
 	assert.Equal(t, uint64(30), oldestKept)
@@ -331,7 +200,7 @@ func TestPruneBlockDataUpto(t *testing.T) {
 	}
 
 	testutils.WithBatch(t, database, func(batch db.Batch) error {
-		return PruneBlockDataUpto(batch, endExclusive)
+		return pruner.PruneBlockDataUpto(batch, endExclusive)
 	})
 
 	// Headers below the lag floor: gone.
